Revert "Organize unit tests into relevant folder"
This reverts commit 5b8e9d72fdc8b5a9c589e11abb87438ad295d51c.
